TP can run multiple functions (F)
Since we already have a test case for a single function the previous
tests are redundant.

Note that we could keep them and simply wrap the function in a list
(so we provide a unified interface that always takes a list of functions
and a list of parameters for the pipeline).